requirements.txt
# Data Collection
tweepy==4.14.0
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.2

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Spark
pyspark==3.4.1
delta-spark==2.4.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# Storage
boto3==1.29.7
minio==7.2.0
s3fs==2023.12.0

# Analytics
duckdb==0.9.2

# Visualization
streamlit==1.28.2
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.3

# NLP (for text processing)
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Future: Vector DB and RAG
# milvus==2.3.3
# langchain==0.0.340
# sentence-transformers==2.2.2
# openai==1.3.7
# anthropic==0.7.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.0
click==8.1.7

# Logging and Monitoring
loguru==0.7.2

# Development Tools
pytest==7.4.3
black==23.12.0
flake8==6.1.0
ipython==8.18.1
jupyter==1.0.0
